separating chart services, adding chart names instead of statics keys. adding ids in table heads
